Add cast to int to silence warning on msvc.

+++ generic/tkFont.c
@@ -10,7 +10,7 @@
  * See the file "license.terms" for information on usage and redistribution of
  * this file, and for a DISCLAIMER OF ALL WARRANTIES.
  *
- * RCS: @(#) $Id: tkFont.c,v 1.59 2009/10/10 17:42:50 dkf Exp $
+ * RCS: @(#) $Id: tkFont.c,v 1.60 2009/10/21 21:22:26 patthoyts Exp $
  */
 
 #include "tkInt.h"
@@ -2378,7 +2378,8 @@ TkDrawAngledTextLayout(
 	    dy = -sinA * (chunkPtr->x + drawX) + cosA * (chunkPtr->y);
 	    if (angle == 0.0) {
 		Tk_DrawChars(display, drawable, gc, layoutPtr->tkfont,
-			firstByte, lastByte - firstByte, x + dx, y + dy);
+			firstByte, lastByte - firstByte,
+			(int)(x + dx), (int)(y + dy));
 	    } else {
 		TkpDrawAngledChars(display, drawable, gc, layoutPtr->tkfont,
 			firstByte, lastByte - firstByte, x+dx, y+dy, angle);
